requirements.txt
# ==== Core Backend Framework ====
Flask==3.0.3
Werkzeug==3.0.2
python-dotenv==1.0.1

# ==== Machine Learning / AI ====
tensorflow-cpu==2.17.0
joblib==1.4.2
scikit-learn==1.6.1
numpy==1.26.4
pandas==2.2.3

# ==== Image Processing ====
opencv-python==4.10.0.84
Pillow==10.4.0

# ==== Visualization (optional but useful for debugging) ====
matplotlib==3.9.2

# ==== External API Calls ====
requests==2.32.3

# ==== JSON Handling ====
simplejson==3.19.2

# ==== For production compatibility ====
gunicorn==23.0.0
